feat: move dropdown mui to ui directory
