+++ src/components/molecules/LeftSidePanelSettings.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Button, MenuItem, TextField, Typography } from "@mui/material";
-import { Dropdown } from "../atoms/DropdownMui";
+import { Dropdown } from "../ui/DropdownMui";
 import { RunSettings } from "./RunSettings";
 import { useSocketStore } from "../../context/socket";
